+++ src/NuGet.Updater/Extensions/NuGetPackageExtensions.cs
@@ -3,60 +3,39 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Updater.Entities;
+using Uno.Extensions;
 
 namespace NuGet.Updater.Extensions
 {
 	public static class NuGetPackageExtensions
 	{
-		public static async Task<FeedNuGetVersion> GetLatestVersion(
+		public static async Task<UpdaterVersion> GetLatestVersion(
 			this NuGetPackage package,
 			CancellationToken ct,
 			UpdaterParameters parameters
 		)
 		{
-			var specialVersion = parameters.TargetVersion;
+			var versions = await package.GetVersions(ct);
 
-			if (parameters.ShouldKeepPackageAtLatestDev(package.PackageId))
-			{
-				specialVersion = "dev";
-			}
+			var versionsPerTarget = versions
+				.OrderByDescending(v => v.Version)
+				.GroupBy(version => parameters.TargetVersions.FirstOrDefault(t => version.IsMatchingVersion(t, parameters.Strict)))
+				.Where(g => g.Key.HasValue());
 
-			if (specialVersion == "stable")
-			{
-				specialVersion = "";
-			}
-
-			var versions = (await package.GetVersions(ct)).OrderByDescending(v => v.Version);
-
-			var version = versions
-				.Where(v => v.IsMatchingSpecialVersion(specialVersion, parameters.Strict) && !v.ContainsTag(parameters.TagToExclude))
+			return versionsPerTarget
+				.Select(g => g.FirstOrDefault())
 				.OrderByDescending(v => v.Version)
 				.FirstOrDefault();
-
-			if(parameters.UseStableIfMoreRecent && specialVersion != "")
-			{
-				var stableVersion = versions
-					.Where(v => v.IsMatchingSpecialVersion("", parameters.Strict) && !v.ContainsTag(parameters.TagToExclude))
-					.OrderByDescending(v => v.Version)
-					.FirstOrDefault();
-
-				if (version == null || (stableVersion?.Version.IsGreaterThan(version.Version) ?? false))
-				{
-					return stableVersion;
-				}
-			}
-
-			return version;
 		}
 
-		private static async Task<IEnumerable<FeedNuGetVersion>> GetVersions(this NuGetPackage package, CancellationToken ct)
+		private static async Task<IEnumerable<UpdaterVersion>> GetVersions(this NuGetPackage package, CancellationToken ct)
 		{
-			var versions = new List<FeedNuGetVersion>();
+			var versions = new List<UpdaterVersion>();
 			foreach(var p in package.Packages)
 			{
 				foreach(var v in await p.Value.GetVersionsAsync())
 				{
-					versions.Add(new FeedNuGetVersion(p.Key, v.Version));
+					versions.Add(new UpdaterVersion(p.Key, v.Version));
 				}
 			}