+++ src/NuGet.Updater/Extensions/PackageSourceExtensions.cs
@@ -7,43 +7,26 @@ using NuGet.Common;
 using NuGet.Configuration;
 using NuGet.Protocol;
 using NuGet.Protocol.Core.Types;
+using NuGet.Updater.Log;
 
 namespace NuGet.Updater.Extensions
 {
 	public static class PackageSourceExtensions
 	{
-		public static async Task<NuGetPackage[]> SearchPackages(this PackageSource source, CancellationToken ct, Action<string> logAction, string searchTerm = "")
+		public static async Task<NuGetPackage[]> SearchPackages(this PackageSource source, CancellationToken ct, string searchTerm = "", Logger log = null)
 		{
 			var settings = Settings.LoadDefaultSettings(null);
 			var repositoryProvider = new SourceRepositoryProvider(settings, Repository.Provider.GetCoreV3());
 
 			var repository = repositoryProvider.CreateRepository(source, FeedType.HttpV3);
 
-			logAction($"Pulling NuGet packages from {source.SourceUri}");
+			log?.Write($"Pulling packages from {source.SourceUri}");
 
 			var searchResource = repository.GetResource<PackageSearchResource>();
 
 			var packages = (await searchResource.SearchAsync(searchTerm, new SearchFilter(true, SearchFilterType.IsAbsoluteLatestVersion), skip: 0, take: 1000, log: new NullLogger(), cancellationToken: ct)).ToArray();
 
-			logAction($"Found {packages.Length} packages");
-
-			return source.ToNuGetPackages(packages);
-		}
-
-		public static async Task<NuGetPackage[]> ListPackages(this PackageSource source, CancellationToken ct, Action<string> logAction, string searchTerm = "")
-		{
-			var settings = Settings.LoadDefaultSettings(null);
-			var repositoryProvider = new SourceRepositoryProvider(settings, Repository.Provider.GetCoreV3());
-
-			var repository = repositoryProvider.CreateRepository(source, FeedType.HttpV3);
-
-			logAction($"Pulling NuGet packages from {source.SourceUri}");
-
-			var listResource = repository.GetResource<ListResource>();
-
-			var packages = await (await listResource.ListAsync(searchTerm, prerelease: true, allVersions: false, includeDelisted: false, log: new NullLogger(), token: ct)).ToArray();
-
-			logAction($"Found {packages.Length} packages");
+			log?.Write($"Found {packages.Length} packages");
 
 			return source.ToNuGetPackages(packages);
 		}