+++ src/NuGet.Updater/NuGetUpdater.cs
@@ -6,14 +6,17 @@ using System.Threading.Tasks;
 using NuGet.Updater.Entities;
 using NuGet.Updater.Extensions;
 using NuGet.Updater.Helpers;
-using NuGet.Configuration;
 using NuGet.Versioning;
+using Uno.Extensions;
+using NuGet.Updater.Log;
 
 #if UAP
 using XmlDocument = Windows.Data.Xml.Dom.XmlDocument;
 #else
 using XmlDocument = System.Xml.XmlDocument;
 #endif
+
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("NuGet.Updater.Tests")]
 
 namespace NuGet.Updater
 {
@@ -22,23 +25,24 @@ namespace NuGet.Updater
 	/// </summary>
 	public partial class NuGetUpdater
 	{
-		private static readonly PackageSource NuGetOrgSource = new PackageSource("https://api.nuget.org/v3/index.json");
-
 		private readonly UpdaterParameters _parameters;
 		private readonly Logger _log;
+		private readonly IUpdaterSource[] _packageSources;
 
-		private NuGetUpdater(UpdaterParameters parameters, Logger log)
+		internal NuGetUpdater(UpdaterParameters parameters, IUpdaterSource[] packageSources, Logger log)
 		{
+			//TODO validate parameters
 			_parameters = parameters;
 			_log = log;
+			_packageSources = packageSources;
 		}
 
-		private async Task<bool> UpdatePackages(CancellationToken ct)
+		internal async Task<bool> UpdatePackages(CancellationToken ct)
 		{
 			_log.Clear();
 
 			var packages = await GetPackages(ct);
-			var targetFiles = await GetTargetFiles(ct);
+			var targetFiles = await FileHelper.GetTargetFiles(ct, _parameters.UpdateTarget, _parameters.SolutionRoot, _log);
 
 			foreach(var package in packages.Where(p => _parameters.ShouldUpdatePackage(p)))
 			{
@@ -46,11 +50,11 @@ namespace NuGet.Updater
 
 				if(latest == null)
 				{
-					_log.Write($"Skipping [{package.PackageId}]: no {_parameters.TargetVersion} version found.");
+					_log.Write($"Skipping [{package.PackageId}]: no {string.Join(" or ", _parameters.TargetVersions)} version found.");
 					continue;
 				}
 
-				_log.Write($"Latest {_parameters.TargetVersion} version for [{package.PackageId}] is [{latest.Version}]");
+				_log.Write($"Latest matching version for [{package.PackageId}] is [{latest.Version}]");
 
 				foreach(var files in targetFiles)
 				{
@@ -84,100 +88,19 @@ namespace NuGet.Updater
 
 		private async Task<NuGetPackage[]> GetPackages(CancellationToken ct)
 		{
-			//Using search instead of list because the latter forces the v2 api
-			var packages = await _parameters.GetFeedPackageSource().SearchPackages(ct, _log.Write);
+			var packagesPerSource = await Task.WhenAll(_packageSources.Select(s => s.GetPackages(ct)));
 
-			if(_parameters.IncludeNuGetOrg)
-			{
-				//Using search instead of list because the latter forces the v2 api
-				packages = packages
-					.Concat(await NuGetOrgSource.SearchPackages(ct, _log.Write, searchTerm: $"owner:{_parameters.PublickPackageOwner}"))
-					.GroupBy(p => p.PackageId)
-					.Select(g => new NuGetPackage(g.Key, g.ToArray()))
-					.ToArray();
-			}
-
-			return packages;
-		}
-
-		private async Task<Dictionary<UpdateTarget, Dictionary<string, XmlDocument>>> GetTargetFiles(CancellationToken ct)
-		{
-			var targetFiles = new Dictionary<UpdateTarget, Dictionary<string, XmlDocument>>();
-
-			var updateTarget = new[]
-			{
-				UpdateTarget.Nuspec,
-				UpdateTarget.PackageReference,
-				UpdateTarget.ProjectJson,
-				UpdateTarget.DirectoryProps,
-				UpdateTarget.DirectoryTargets,
-			};
-
-			foreach(var target in updateTarget)
-			{
-				if(_parameters.HasUpdateTarget(target))
-				{
-					targetFiles.Add(target, await GetFilesForTarget(ct, target));
-				}
-			}
-
-			return targetFiles;
-		}
-
-		private async Task<Dictionary<string, XmlDocument>> GetFilesForTarget(CancellationToken ct, UpdateTarget target)
-		{
-			string extensionFilter = null, nameFilter = null;
-
-			switch(target)
-			{
-				case UpdateTarget.Nuspec:
-					extensionFilter = ".nuspec";
-					break;
-
-				case UpdateTarget.ProjectJson:
-					nameFilter = "project.json";
-					break;
-
-				case UpdateTarget.PackageReference:
-					extensionFilter = ".csproj";
-					break;
-
-				case UpdateTarget.DirectoryTargets:
-					nameFilter = "Directory.Build.targets";
-					break;
-
-				case UpdateTarget.DirectoryProps:
-					nameFilter = "Directory.Build.props";
-					break;
-
-				default:
-					break;
-			}
-
-			if(extensionFilter == null && nameFilter == null)
-			{
-				return new Dictionary<string, XmlDocument>();
-			}
-
-			_log.Write($"Retrieving {nameFilter ?? extensionFilter} files");
-
-			var files = await FileHelper.GetFiles(ct, _parameters.SolutionRoot, extensionFilter, nameFilter);
-
-			_log.Write($"Found {files.Length} {nameFilter ?? extensionFilter} file(s)");
-
-			if(target == UpdateTarget.ProjectJson)
-			{
-				return files.ToDictionary(f => f, f => default(XmlDocument));
-			}
-
-			return (await Task.WhenAll(files.Select(f => f.GetDocument(ct))))
-				.ToDictionary(p => p.Key, p => p.Value);
+			return packagesPerSource
+				.SelectMany(x => x)
+				.GroupBy(p => p.PackageId)
+				.Select(g => new NuGetPackage(g.Key, g.ToArray()))
+				.ToArray();
 		}
 
 		private static async Task<UpdateOperation[]> UpdateNuSpecs(
 			CancellationToken ct,
 			string packageId,
-			FeedNuGetVersion version,
+			UpdaterVersion version,
 			Dictionary<string, XmlDocument> nuspecDocuments,
 			bool isDowngradeAllowed
 		)
@@ -191,7 +114,7 @@ namespace NuGet.Updater
 
 				var updates = xmlDocument.UpdateNuSpecVersions(packageId, version, path, isDowngradeAllowed);
 
-				if(updates.Any(u => u.ShouldProceed))
+				if(updates.Any(u => u.IsUpdate))
 				{
 					await xmlDocument.Save(ct, path);
 				}
@@ -205,7 +128,7 @@ namespace NuGet.Updater
 		private static async Task<UpdateOperation[]> UpdateProjects(
 			CancellationToken ct,
 			string packageId,
-			FeedNuGetVersion version,
+			UpdaterVersion version,
 			Dictionary<string, XmlDocument> projectDocuments,
 			bool isDowngradeAllowed
 		)
@@ -219,7 +142,7 @@ namespace NuGet.Updater
 
 				var updates = xmlDocument.UpdateProjectReferenceVersions(packageId, version, path, isDowngradeAllowed);
 
-				if(updates.Any(u => u.ShouldProceed))
+				if(updates.Any(u => u.IsUpdate))
 				{
 					await xmlDocument.Save(ct, path);
 				}
@@ -233,7 +156,7 @@ namespace NuGet.Updater
 		private static async Task<UpdateOperation[]> UpdateProjectJson(
 			CancellationToken ct,
 			string packageName,
-			FeedNuGetVersion latestVersion,
+			UpdaterVersion latestVersion,
 			string[] jsonFiles,
 			bool isDowngradeAllowed
 		)
@@ -255,7 +178,7 @@ namespace NuGet.Updater
 
 					var operation = new UpdateOperation(isDowngradeAllowed, packageName, currentVersion, latestVersion, file);
 
-					if(operation.ShouldProceed)
+					if(operation.IsUpdate)
 					{
 						var newContent = Regex.Replace(
 							fileContent,